fix: download json from precalculated form
