+++ newfrontend/src/app/utils/utils.ts
@@ -4,3 +4,13 @@ export function allTrue(arr: any[]): boolean {
     }
     return true;
 }
+
+export function downloadJson(name: string, data: any) {
+    const dataStr = "data:text/json;charset=utf-8," + encodeURIComponent(JSON.stringify(data));
+    var downloadAnchorNode = document.createElement('a');
+    downloadAnchorNode.setAttribute("href",     dataStr);
+    downloadAnchorNode.setAttribute("download", name + ".json");
+    document.body.appendChild(downloadAnchorNode); // required for firefox
+    downloadAnchorNode.click();
+    downloadAnchorNode.remove();
+  }
